+++ strelets/remove_virtual_chain.go
@@ -2,7 +2,6 @@ package strelets
 
 import (
 	"context"
-	"github.com/orbs-network/boyarin/strelets/adapter"
 )
 
 type RemoveVirtualChainInput struct {
@@ -10,6 +9,5 @@ type RemoveVirtualChainInput struct {
 }
 
 func (s *strelets) RemoveVirtualChain(ctx context.Context, input *RemoveVirtualChainInput) error {
-	serviceName := adapter.GetServiceId(input.VirtualChain.GetContainerName())
-	return s.orchestrator.ServiceRemove(ctx, serviceName)
+	panic("removed")
 }
