+++ strelets/shared_network.go
@@ -2,7 +2,6 @@ package strelets
 
 import (
 	"context"
-	"fmt"
 )
 
 type ProvisionSharedNetworkInput struct {
@@ -10,9 +9,5 @@ type ProvisionSharedNetworkInput struct {
 }
 
 func (s *strelets) ProvisionSharedNetwork(ctx context.Context, input *ProvisionSharedNetworkInput) error {
-	_, err := s.orchestrator.GetOverlayNetwork(ctx, input.Name)
-	if err != nil {
-		return fmt.Errorf("failed to provision shared network %s: %s", input.Name, err)
-	}
-	return nil
+	panic("removed")
 }
