+++ strelets/update_reverse_proxy.go
@@ -2,9 +2,7 @@ package strelets
 
 import (
 	"context"
-	"fmt"
 	"github.com/orbs-network/boyarin/strelets/adapter"
-	"io/ioutil"
 )
 
 type UpdateReverseProxyInput struct {
@@ -15,25 +13,5 @@ type UpdateReverseProxyInput struct {
 }
 
 func (s *strelets) UpdateReverseProxy(ctx context.Context, input *UpdateReverseProxyInput) error {
-	sslEnabled := input.SSLOptions.SSLCertificatePath != "" && input.SSLOptions.SSLPrivateKeyPath != ""
-
-	config := &adapter.ReverseProxyConfig{
-		NginxConfig: getNginxConfig(input.Chains, input.IP, sslEnabled),
-	}
-
-	if sslEnabled {
-		if sslCertificate, err := ioutil.ReadFile(input.SSLOptions.SSLCertificatePath); err != nil {
-			return fmt.Errorf("could not read SSL certificate from %s: %s", input.SSLOptions.SSLCertificatePath, err)
-		} else {
-			config.SSLCertificate = sslCertificate
-		}
-
-		if sslPrivateKey, err := ioutil.ReadFile(input.SSLOptions.SSLPrivateKeyPath); err != nil {
-			return fmt.Errorf("could not read SSL private key from %s: %s", input.SSLOptions.SSLCertificatePath, err)
-		} else {
-			config.SSLPrivateKey = sslPrivateKey
-		}
-	}
-
-	return s.orchestrator.RunReverseProxy(ctx, config)
+	panic("removed")
 }
